# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tracking API",
        "description": "Documentación de la Tracking API",
        "contact": {
            "email": "soporte@tudominio.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/users/{user}/roles": {
            "post": {
                "tags": [
                    "UserRoles"
                ],
                "summary": "Asignar rol a usuario",
                "operationId": "ce35601debe2afb3295a34c4496efed1",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "role": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rol asignado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users/{user}/roles/{role}": {
            "delete": {
                "tags": [
                    "UserRoles"
                ],
                "summary": "Revocar rol de usuario",
                "operationId": "3ab0db0ea3c989ad5b23f7b22988ade4",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "role",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rol revocado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AuthResponse": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 200
                    },
                    "message": {
                        "type": "string",
                        "example": "Autenticación exitosa"
                    },
                    "data": {
                        "properties": {
                            "user": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "johnd"
                                    },
                                    "photoUrl": {
                                        "type": "string",
                                        "example": "https://…"
                                    }
                                },
                                "type": "object"
                            },
                            "access_token": {
                                "type": "string",
                                "example": "jwt.token.here"
                            },
                            "token_type": {
                                "type": "string",
                                "example": "bearer"
                            },
                            "expires_in": {
                                "type": "integer",
                                "example": 3600
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}